+++ src/connection_handler.rs
@@ -1,10 +1,15 @@
+use std::net::IpAddr;
+
 use async_trait::async_trait;
 #[cfg(test)]
 use mockall::automock;
 use ssh_key::Fingerprint;
 use tokio::sync::mpsc;
 
+// Extra data available for HTTP tunneling/aliasing connections.
+#[derive(Clone)]
 pub(crate) struct ConnectionHttpData {
+    // Port to redirect HTTP requests to. If missing, do not redirect.
     pub(crate) redirect_http_to_https_port: Option<u16>,
 }
 
@@ -16,12 +21,12 @@ pub(crate) trait ConnectionHandler<T: Sync> {
     fn log_channel(&self) -> Option<mpsc::UnboundedSender<Vec<u8>>>;
 
     // Return the tunneling channel for this connection.
-    async fn tunneling_channel(&self, ip: &str, port: u16) -> anyhow::Result<T>;
+    async fn tunneling_channel(&self, ip: IpAddr, port: u16) -> anyhow::Result<T>;
 
     // Return the aliasing channel for this connection.
     async fn aliasing_channel<'a>(
         &self,
-        ip: &str,
+        ip: IpAddr,
         port: u16,
         fingerprint: Option<&'a Fingerprint>,
     ) -> anyhow::Result<T>;