+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add User-Agent to login API requests
 - Add `force-https` option for remote forwarding connections.
+- Add `ip-allowlist`/`ip-blocklist` options for remote forwarding connections.
 
 ## 0.3.1 (2024-12-31)